[linux] separate configure, build and install steps
build steps are now:
make configure (if needed)
make
sudo make install

or:
make fullbuild
sudo make install

+++ linux/scripts/build.sh
@@ -1,72 +1,53 @@
 #!/bin/bash
 
 # Build KMFL in the required order: kmflcomp libkmfl ibus-kmfl
-# install each to temp install dir INSTALLDIR which is /tmp/kmfl by default
-#   so that libkmfl and ibus-kmfl can use what they need that have been installed
 
 # It must be run from the keyman/linux directory
-
-# To install the ibus-kmfl xml file and icons to /usr/share so that ibus
-# will use ibus-kmfl you must run SUDOINSTALL="yes" ./build.sh
-# This will install everything else to /usr/local
 
 set -e
 
 BASEDIR=`pwd`
 
-SUDOINSTALL=${SUDOINSTALL:-"no"}
+INSTALLDIR=${INSTALLDIR:-"/usr/local"}
 
-INSTALLDIR=${INSTALLDIR:-"/tmp/kmfl"}
+CONFIGUREONLY=${CONFIGUREONLY:="no"}
+BUILDONLY=${BUILDONLY:="no"}
 
-if [ -f "/usr/share/ibus/component/kmfl.xml" ] && [ "${SUDOINSTALL}" == "yes" ]; then
-	if grep -Fq "/usr/lib/ibus" /usr/share/ibus/component/kmfl.xml
-	then
-		echo "component file is ibus-kmfl package version so move it"
-		echo "run 'SUDOINSTALL=uninstall ./build.sh' to put it back"
-		sudo mv /usr/share/ibus/component/kmfl.xml /usr/share/doc/ibus-kmfl/
-	else
-		echo "component file is local one so overwrite it"
-	fi
+if [[ "${CONFIGUREONLY}" != "no" && "${BUILDONLY}" != "no" ]]; then
+	echo "Only use one of CONFIGUREONLY and BUILDONLY"
+	exit 1
 fi
 
 for proj in kmflcomp libkmfl ibus-kmfl; do
-	rm -rf build-$proj
-	mkdir build-$proj
+	if [ ! -f $proj/configure ]; then
+		echo "$proj is not set up with autoreconf. First run 'make reconf' or 'make devreconf'"
+		exit 1
+	fi
+	if [[ "${BUILDONLY}" == "no" ]]; then
+		rm -rf build-$proj
+	fi
+	mkdir -p build-$proj
 	cd build-$proj
-	if [[ "${SUDOINSTALL}" == "yes" ]]; then
-		echo "doing sudo install of $proj"
-		../$proj/configure CPPFLAGS="-I$/usr/local/include" LDFLAGS="-L/usr/local/lib" --libexecdir=/usr/local/lib/ibus --datadir=/usr/share
+	if [[ "${BUILDONLY}" == "no" ]]; then
+		if [[ "${INSTALLDIR}" == "/tmp/kmfl" ]]; then # don't install ibus-kmfl into ibus
+			../$proj/configure CPPFLAGS="-I../kmflcomp/include -I../libkmfl/include" LDFLAGS="-L`pwd`/../build-kmflcomp/src -L`pwd`/../build-libkmfl/src" --prefix=${INSTALLDIR} --libexecdir=${INSTALLDIR}/lib/ibus
+		else	# install ibus-kmfl into ibus
+			../$proj/configure CPPFLAGS="-I../kmflcomp/include -I../libkmfl/include" LDFLAGS="-L`pwd`/../build-kmflcomp/src -L`pwd`/../build-libkmfl/src" --prefix=${INSTALLDIR} --libexecdir=${INSTALLDIR}/lib/ibus --datadir=/usr/share
+		fi
+	fi
+	if [[ "${CONFIGUREONLY}" == "no" ]]; then
+		if [ ! -f config.h ]; then
+			echo "$proj has not been configured before building. First run 'make configure'"
+			exit 1
+		fi
 		make
-		sudo make install
-	elif [[ "${SUDOINSTALL}" == "uninstall" ]]; then
-		echo "doing sudo uninstall of $proj"
-		../$proj/configure CPPFLAGS="-I$/usr/local/include" LDFLAGS="-L/usr/local/lib" --libexecdir=/usr/local/lib/ibus --datadir=/usr/share
-		sudo make uninstall
-	else
-		echo "doing /tmp install of $proj"
-		../$proj/configure CPPFLAGS="-I${INSTALLDIR}/include" LDFLAGS="-L${INSTALLDIR}/lib" --prefix=${INSTALLDIR} --libexecdir=${INSTALLDIR}/lib/ibus
-		make
-		make install
 	fi
 	cd $BASEDIR
 done
 
-cd keyman-config
-echo "SUDOINSTALL: ${SUDOINSTALL}"
-if [[ "${SUDOINSTALL}" == "yes" ]]; then
-	echo "doing sudo install of keyman-config"
-	make install
-elif [[ "${SUDOINSTALL}" == "uninstall" ]]; then
-	echo "doing sudo uninstall of keyman-config"
-	make uninstall
-else
-	echo "doing /tmp install of keyman-config"
-	make install-temp
+if [[ "${CONFIGUREONLY}" == "no" ]]; then
+	cd keyman-config
+	make clean
+	make
+	cd $BASEDIR
 fi
-cd $BASEDIR
-
-if [ -f "/usr/share/doc/ibus-kmfl/kmfl.xml" ] && [ "${SUDOINSTALL}" == "uninstall" ]; then
-	echo "putting ibus-kmfl package component file back"
-	sudo mv /usr/share/doc/ibus-kmfl/kmfl.xml /usr/share/ibus/component/
-fi
-
